pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "devpulse-cli"
version = "0.1.0"
description = "A CLI for monitoring GitHub PRs, CI pipelines, and team velocity"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Arun Kumar", email = "arunch.pluto@gmail.com"}]
dependencies = [
    "click>=8.1",
    "httpx>=0.25",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "respx>=0.20"]

[project.scripts]
devpulse = "devpulse.cli:main"

[project.urls]
Homepage = "https://github.com/Arunchunchukumar/devpulse-cli"
Issues = "https://github.com/Arunchunchukumar/devpulse-cli/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
